fix: garante que a variável será calculada como int (#1)
(cherry picked from commit 47decef377a8de0c7b8f7f0aacc6d852291c770c)

+++ src/Bancos/BoletoSicoob.php
@@ -50,7 +50,7 @@ class BoletoSicoob implements NossoNumeroInterface
                 $constante = 7;
                 $cont = 0;
             }
-            $calculoDv = $calculoDv + (substr($sequencia, $num, 1) * $constante);
+            $calculoDv = $calculoDv + ((int) substr($sequencia, $num, 1) * $constante);
         }
 
         $Resto = $calculoDv % 11;
